BOOST-700 reorganize folder structures and refactors code for integration tests
