+++ packages/framework-integration-tests/integration/cli/setup.ts
@@ -1,6 +1,11 @@
 import * as path from 'path'
 import util = require('util')
+import { removeFile } from '../helper/fileHelper'
+import { CLI_ENTITY_INTEGRATION_TEST_FILES } from './cli.entity.integration'
+
 const exec = util.promisify(require('child_process').exec)
+
+const testFiles: Array<string> = [...CLI_ENTITY_INTEGRATION_TEST_FILES]
 
 before(async () => {
   const integrationTestsPackageRoot = path.dirname(__dirname)
@@ -9,7 +14,20 @@ before(async () => {
   await exec('lerna bootstrap')
   await exec('lerna clean --yes')
   await exec('lerna run clean')
-  await exec('lerna run compile')
 
   process.chdir('..')
+
+  try {
+    await Promise.all(testFiles.map((file: string) => removeFile(path.join(path.dirname(__dirname), '..', file))))
+  } catch (e) {
+    // error whilst deleting files
+  }
+})
+
+after(async () => {
+  try {
+    await exec('lerna run compile')
+  } finally {
+    await Promise.all(testFiles.map((file: string) => removeFile(path.join(path.dirname(__dirname), '..', file))))
+  }
 })
