+++ packages/framework-integration-tests/integration/fixtures/entities/PostWithReducer.ts
