+++ packages/framework-integration-tests/integration/fixtures/events/PostCreated.ts
