+++ packages/framework-integration-tests/integration/providers/aws/.mocharc.yml
@@ -7,6 +7,6 @@ recursive: true
 reporter: 'spec'
 timeout: 600000
 file:
-  - './integration/aws/setup.ts'
+  - './integration/providers/aws/setup.ts'
 full-trace: true
 # bail: true -- This option makes the test suite stop after the first failure. Might make sense to enable once tests are stable
