+++ packages/framework-integration-tests/integration/providers/aws/auth.integration.ts
