+++ packages/framework-integration-tests/integration/providers/aws/stack.integration.ts
