+++ packages/framework-integration-tests/integration/providers/aws/utils.ts
