+++ packages/framework-integration-tests/package.json
@@ -43,7 +43,7 @@
     "compile": "tsc -b tsconfig.json",
     "clean": "rimraf ./dist tsconfig.tsbuildinfo",
     "integration": "yarn integration/aws && yarn integration/local && yarn integration/cli",
-    "integration/aws": "mocha --forbid-only --config \"integration/aws/.mocharc.yml\" \"integration/aws/**/*.integration.ts\"",
+    "integration/aws": "mocha --forbid-only --config \"integration/providers/aws/.mocharc.yml\" \"integration/providers/aws/**/*.integration.ts\"",
     "integration/local": "mocha --forbid-only --config \"integration/local/.mocharc.yml\" \"integration/local/**/*.integration.ts\"",
     "integration/cli": "mocha --forbid-only --config \"integration/cli/.mocharc.yml\" \"integration/cli/**/*.integration.ts\""
   },
